refactor search functionality into package
