--- cvesearch/scripts/__init__.py
@@ -1,28 +1,0 @@
-import click
-from .search import search as __search
-from .cpe import search_cpe
-from .irc import search_irc
-from .xmpp import search_xmpp
-from .fulltext import search_fulltext
-from . import db
-
-
-@click.group()
-@click.pass_context
-def _search(ctx):
-    if ctx.invoked_subcommand is (None | 'search'):
-        return __search(ctx)
-    elif ctx.invoked_subcommand is (search_cpe | 'cpe'):
-        return search_cpe(ctx)
-    elif ctx.invoked_subcommand is (search_irc | 'irc'):
-        return search_irc(ctx)
-    elif ctx.invoked_subcommand is (search_xmpp | 'xmpp'):
-        return search_xmpp(ctx)
-    elif ctx.invoked_subcommand is (search_fulltext | 'fulltext'):
-        return search_fulltext(ctx)
-    ...
-
-
-Search = _search
-
-__all__ = [Search, db]
