+++ cvesearch/search/__init__.py
@@ -10,22 +10,27 @@
 
 # Imports
 import click
+
+from .cpe import search_cpe
+from .irc import search_irc
+from .xmpp import search_xmpp
+from .fulltext import search_fulltext
+
 from .. import DatabaseLayer, Query, output
-from . import _search
 
 
-@_search.command()
+@click.command()
 @click.pass_context
-@_search.option('-p', '--product', type=click.STRING)
+@click.option('-p', '--product', type=click.STRING)
 # @click.argument('-f', '--fulltext', help='full text search in cve summary')
-@_search.argument('-c', '--cve', type=click.STRING, multi=True)
-@_search.option('-o', '--output', type=click.STRING, choices=['csv', 'html', 'json', 'xml', 'cveid', 'asciidoc'])
+@click.argument('-c', '--cve', type=click.STRING, multi=True)
+@click.option('-o', '--output', type=click.STRING, choices=['csv', 'html', 'json', 'xml', 'cveid', 'asciidoc'])
 # @click.argument('-l', '--descending', is_flag=True, help='sort in descending mode')
-@_search.option('-n', '--cpe', type=click.BOOL)
-@_search.option('-r', '--ranking', type=click.BOOL)
-@_search.option('-a', '--capec', type=click.BOOL)
+@click.option('-n', '--cpe', type=click.BOOL)
+@click.option('-r', '--ranking', type=click.BOOL)
+@click.option('-a', '--capec', type=click.BOOL)
 # @click.argument('-v', '--vendor', help='vendor name to lookup in reference urls')
-@_search.option('--api', type=click.STRING)
+@click.option('--api', type=click.STRING)
 def search(product, cve, frmt, cpe, rank, capec, api):
     """Search for vulnerabilities in the Nat'l Vulnerability Database(https://nvd.nist.gov)
 
@@ -78,6 +83,23 @@ def search(product, cve, frmt, cpe, rank, capec, api):
 
     # Print output
     return _search.echo(output(items, frmt, **kwargs))
+
+
+@search.group()
+@click.pass_context
+def _search(ctx):
+    if ctx.invoked_subcommand is (None | 'search'):
+        return search(ctx)
+    elif ctx.invoked_subcommand is (search_cpe | 'cpe'):
+        return search_cpe(ctx)
+    elif ctx.invoked_subcommand is (search_irc | 'irc'):
+        return search_irc(ctx)
+    elif ctx.invoked_subcommand is (search_xmpp | 'xmpp'):
+        return search_xmpp(ctx)
+    elif ctx.invoked_subcommand is (search_fulltext | 'fulltext'):
+        return search_fulltext(ctx)
+    else:
+        ...
 
 
 if __name__ == '__main__':
